adding scaffolding for operator
Step 1 of getting this working is to create an initial design or
scaffolding that defines the structure of the operators. Here we have
a FluxJob -> controlled by the Flux controller, which will expose a
custom resource (the user yaml file) that has the container image and
command. The more complex setup (e.g., configs and statefulset) are
managed by the FluxSetup, a more admin/internal controller. For that
one, right now a lot of the logic is just represented as strings in
the API, and as this is developed these will be flushed out into
more structs/types proper. This is just the first outline/structure
for how the operator will look! I will push logical sets of changes
as I work on them, and bring up for review/discussion when I have
something mostly working.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>