consolidating logic to create config maps
both the config map for the broker and the etc-hosts are similar
in that they have hostfile for data, so we should not need separate
(and redundant) functions to create them. I started creating the secret,
but I am not sure how it works because there are instructions to uncomment
the CERTMANAGER but also a declaration that it is not supported.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>