start of refactor to add flux job and minicluster
we conceptually want a third CRD that can hold a state of a mini cluster.
with this addition, the user will submit a FluxJob CR, and then it will
generate a unique identifier, look for a matching mini-cluster, and if
not found create it. Then it will keep checking until the mini-cluster
status goes from PENDING waiting for resources -> READY, and then run
the jobs. On the other side, the FluxSetup will be watching the mini-cluster
resource, and will clean them up and change status to DONE/FINISHED when
the pods are done running.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>