start of work to refactor to have shared manager
I could not get a third CRD to work (a minicluster) so instead I am trying
the same strategy as kueue - having a manager that can hold a state of queues
and then a fluxjob that adds to it, and a separate fluxsetup that (will) control
assigning jobs to variables. I have not really figured it out all yet, but
I have the basics running and watchers working (for the first time) so I think
this is an OK direction.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>