Design 3: MiniCluster CRD (#13)
* adding MiniCluster CRD design

this is a simplified design that has one CRD, a minicluster, which
is a set of configs and a batch job to run pods (containers), and
they can all be associated with the single CRD to better track state
to coincide with cluster state!

* switch to using indexed job

and prebuilt demo container with lammps and mpi! This still
needs tweaking, but since it is our first working example I think
we should call this a good point to merge into main, and then
address the other issues (re volumes). I am also adding an exposed
ability to customize the entrypoint and ask for alwaysPull since
I found I wanted that just now.

* remove pr build for demo container

we really only need to build on demand (or changes to test)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>