Getting curve.cert setup (#15)
* WIP and start of work for networking!

This includes adding a parameter to the job to ask for
creating a service. I did not get to the hosts debugging yet
because the broker does not like my curve cert! I have the
correct permission and (likely tomorrow) will debug
what it is supposed to look like vs. what I have. I also
cleaned up the volunes so a minicluster volume is under
the same namespace of the job.

* adding generation of zmq curve.cert via init container

instead of requiring a dependncy on the host (for the operator
to generate) or for the operator user to generate it, we can
create an init container alongisde the job worker container that
uses flux keygen to generate the /mnt/curve/curve.cert first!
This obviously needs testing but the broker seems to be happy

Signed-off-by: vsoch <vsoch@users.noreply.github.com>