adding update to/etc/hosts
this is a bit of a hack that adds a script wrapper to the pod start,
and the wrapper waits until it sees a file populated with ip addresses
(or more specifically, echos to update /etc/hosts.). I think we can
do this because when the pod is re-created, the ip address does not
change! And what is happening while it is waiting is that a config
map is updated with the (now known) ip addresses. This seems to
allow the /etc/hosts to be correctly populated (determined by ping
working) and I think next I need to debug why the broker still
thinks it is waiting.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>