separating main node (to generate cert and start) from workers
my flux still is not connecting "Unable to connect to Flux"
so I need to debug this. But I (?)think it is more correct that
only one of the nodes is running the start command and generating
the certificate. Technically if this node knows that it can use
the other ones we should not need to run it multiple times.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>