job is working for the first time!
This likely is not perfect, but at least for the first time I see
the mpirun running and complete, and then the pods complete. I
think next we should discuss what should be the proper start
command for the main/worker nodes, how we can print better verbose
debugging output (possibly exposed by a variable), and have some
solid evidence the node communication is successful (or is the job
running that evidence?

Signed-off-by: vsoch <vsoch@users.noreply.github.com>