testing updated container build
we need to ensure that the lmp executable is available to
the flux user, and not just root (it was by default installed to
root home). We also should demonstrate creating the flux user
and ensuring that permissions are set so he/she/it/they has
access to the files we plan to exec. These small tweaks (per my
local testing) should get the whole demo running as we expect.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>