Updated entrypoints so pods run and complete (#23)
* tweaking entrypoint so workers just start
* adding maximum runtime (seconds) and ensuring all ranks are required

This update will add a maximum time in seconds, which absolutely
cannot be 0 or the job will not start, so instead the default is set
to be a year. This also changes the broker settings so that all
ranks are required to be online before starting. We eventually might
want to expose this as a variable if there are jobs that have the
quality that a worker might die or go offline, or some other
edge case. With this change, the rank 0 waits and they all clean
up and reach completed state. yay!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>