build: testing docker build and bundle
I am not sure what we will want for a more production
deployment (and I need to read up on what a catalog is)
but the base docker image and bundle to install directly
seems like a good idea! This workflow will
add in those two things, and when we know how we want to
deploy this we can tweak further. I also need to poke
around and see what others are doing.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>