fixing networking hack!
This fix will remove the need to restart pods with an /etc/hosts -
the core problem was that developing with MiniKube requires enabling
plugins for DNS. Once those are enabled, we can make a service
alongside the pods for the indexed job, give that name to the
subdomain of the podspec, and then ensure that the flux broker config
has the full hostname added to it. Note that flux is not able to take
a fully qualified hostname that does not match the output of $(hostname).
It also looks like changing the $(hostname) output to FQDN triggers
a bug in flux. So the fix is to include the hard coded name/prefix
in the config instead (that works!). This fix also means that the
minikube service expose command works, but I still prefer being
able to map to port 5000 on localhost.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>