add support for flux option flags
this adds support for fluxOptionFlags to be defined on the level
of the cluster. We check if the variable is defined, and if not,
we do not attempt to export in the environment, as the base container
might already specify a preference. To test this fully I am waiting
for the osu benchmarks container to build, and then I can test it
here with an example workflow.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>