preparing to test state check
when running in the cloud, it seems that the flux worker will
start when it is not ready, and there is no error so nothing
runs. This is not an ideal solution, but here we force the
pods to wait until they are flagged ready (via a filesystem
marker) and then allow the commands to run. I think a better
solution would be to immediately mark the main broker as
ready and start flux, but I need to try that next.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>