Add/manual state check (#55)
* start of work to remove RWX dependency
* tweak developer docs to describe curve generation
* remove debug yaml functionality, is not super helpful
* ensure we have pods/log permission

currently we are requiring a read/write volume shared by pods that
will allow generating a curve.cert on the fly (by the main broker)
and then copying to the other notes. This (I have learned) is a generally
bad strategy. Instead, we create a one-off pod using the same flux runner
container that can generate the certificate, we retrieve that via the logs,
and then we can write it to the MiniCluster via a config map (like
everything else!)

* preparing to test state check
* add gcp docs and remove extra (not used) functionality to exec
* add gcp support / images / docs

when running in the cloud, it seems that the flux worker will
start when it is not ready, and there is no error so nothing
runs. This is not an ideal solution, but here we force the
pods to wait until they are flagged ready (via a filesystem
marker) and then allow the commands to run. I think a better
solution would be to immediately mark the main broker as
ready and start flux, but I need to try that next.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>