fixing bug with uvicorn web ui
We currently had a bug that using the UI meant we could not submit
jobs that required > 1 process to flux. The reason is because the
uvicorn server needs to be handed as many workers as the size that is
needed, and then that should still get passed to flux and submit
appropriately. I am also moving deployment docs into their own
section (e.g., for cloud) since it is not technically development

Signed-off-by: vsoch <vsoch@users.noreply.github.com>