early work to add resources (needs to be tested) (#61)
This was tricker than I anticipated because CRDs do not
easily handle interfaces. The solution is to use an unstructured.Unstructed
type, which then we can write a custom DeepCopy function for, and use
that object as a map[string]interface{} in the controller (which is
what we originally wanted). We will use this branch to test and work
with flux-cloud requesting a special network

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>