support for single services to run alongside minicluster (#132)
* support for single services

If we want to run a singularity container on all nodes of the cluster, we either
need to provision volumes (harder) or pull the container to each node. The latter,
of course, is a really bad idea - we do not want to strain networking to the outside
internet or a registry! A solution is to implement a single service, not a sidecar
in that the service runs alongside each pod (N times) but rather a single service for
the entire cluster. In this case we create a registry, and then we can (once) pull the
container with ORAS to a Singularity image, and then also use oras to push to the OIC
oops OCI compatible registry. Then all the nodes can pull that same container from our
local registry to their systems. In a way it is acting as a pull through cache!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>