add pod completion
we have also figured out some of the issue with flux taking
longer to start now. When we had the certificate generator pod,
this must have somehow kicked off earlier creation of the headless
service so it was ready by the time the indexed job came out. Now
with the redesign, the first attempts to connect by zeromq fail,
and there is some kind of exponential backoff making the start
very slow. When we manually adjust this timeout/wait in Flux
to make it hard coded and shorter, this bug should be resolved.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>