adding tbon connect timeout
This will resolve an issue that zeromq by default
has a timeout on retry with exponential backoff. Although
there is still a remaining issue in Kubernetes that adding
a service pod seems to improve network readiness, for now
this addition of the zeromq timeout will reduce the startup
times from grossly large to acceptably so.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>