experimental work to add flub
I think we are close - I have added the other MiniCluster host names to a shared
service, and the appropriate size and broker options. The nodes register as down,
so likely I have an issue with how the names are defined in the resources or broker
option, or we need to start one of the clusters with a different tree structure
or similar.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>