cleanup client commands
the design I have seen  more often is to use the controller client directly
on the reconciler. We cannot do that directly for Create since we have
events (that require then name) but we can wrap the command and use r.New
instead (and all others, r.Patch, r.Status, R.Get appear to work).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>