add support for custom network name and nginx example
in preparing to test bursting, we need a means of communication
between some external entity that can interact with the rest of
the cluster. While this likely is not a final solution over TLS,
it might be a good first pass at some kind of hack that can allow
communication between two clusters.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>