add pod affinity and antiffinity rules
Add soft rules for the Job so Pods prefer to be scheduled on the same
zone and not to be scheduled on the same node.

This will improve the performance as pods in the same node will still
resources from each other, and pods in different zone will have a
performance penalty since networking is most likely to be worse.

Signed-off-by: Antonio Ojea <aojea@google.com>