update docs for storage and remove singularity examples
We no longer need to have a container built with flux inside, so using
singularity further inside of an application container is not a likely use
case. I am removing them for now. I also tested the filestore setup on GCP
and for now am removing the fusion setup since it is not open source.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>