bug: pre->command block is run twice
Problem: the Commands.Init was added to replace Commands.Pre, and Commands.Pre
was moved lower. But the original one was not. This means that if the user
defines commands in Commands.Pre, they are run twice.
Solution: remove the top one, as I think I originally intended to do

Signed-off-by: vsoch <vsoch@users.noreply.github.com>