bug: race condition for service name
A service name (defaulting to flux-service) that is shared by two
MiniClusters can lead to a race condition, where one cluster is cleaning
up (but the service still exists, associated with the deleting job) and then
the new cluster coming up looks for it (and finds it) and does not recreate.
In practice the network will fail to bootstrap. I believe we have seen this
issue on the rabbit nodes, and I just saw it in scaling experiments.
This change will test if the issue is resolved.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>